functions/produce_delta_event/requirements.txt
google-cloud-pubsub==2.1.0
google-cloud-storage==1.31.2
pandas==0.24.2
gobits==0.0.6
Brotli==1.0.9
xlrd==1.2.0
XlsxWriter==1.1.8
